// 05032026/sce-front/src/app/model/eleccionResponseBean.ts
export class EleccionResponseBean {
  id:string;
  ccodigo: string;
  cnombre: string;
  nombre?:string;
  codigo?:string;
  nactivo: string;
  rangoInicial: string;
  rangoFinal: string;
  digitoChequeoAE: string;
  digitoChequeoAIS: string;
  digitoCequeoError: string;
  preferencial: number;
}